+++ jax/_src/pallas/mosaic_gpu/lowering.py
@@ -2888,6 +2888,20 @@ def _log_lowering_rule(ctx: LoweringRuleContext, x, accuracy):
   return math_dialect.log(_ensure_ir_value(x, x_aval.dtype), fastmath=fastmath)
 
 
+@register_lowering_rule(lax.log2_p, mgpu.LoweringSemantics.Lane)
+@register_lowering_rule(lax.log2_p, mgpu.LoweringSemantics.Warpgroup)
+def _log2_lowering_rule(ctx: LoweringRuleContext, x, accuracy):
+  if accuracy is not None:
+    raise NotImplementedError("Not implemented: accuracy")
+  [x_aval] = ctx.avals_in
+  if ctx.module_ctx.lowering_semantics == mgpu.LoweringSemantics.Lane:
+    return _ensure_fa(x, x_aval.dtype).log2(approx=ctx.module_ctx.approx_math)
+  fastmath = (
+      arith_dialect.FastMathFlags.afn if ctx.module_ctx.approx_math else None
+  )
+  return math_dialect.log2(_ensure_ir_value(x, x_aval.dtype), fastmath=fastmath)
+
+
 @register_lowering_rule(lax.abs_p, mgpu.LoweringSemantics.Lane)
 def _abs_lowering_rule(ctx: LoweringRuleContext, x):
   [x_aval] = ctx.avals_in