+++ jax/experimental/mosaic/gpu/layout_inference.py
@@ -762,6 +762,7 @@ for _op in [
     mlir_math.SinOp,
     mlir_math.CosOp,
     mlir_math.LogOp,
+    mlir_math.Log2Op,
     mlir_math.RsqrtOp,
     mlir_math.TanhOp,
     mlir_math.AbsFOp,