+++ jax/extend/core/primitives.py
@@ -90,6 +90,7 @@ from jax._src.lax.lax import (
   le_p as le_p,
   le_to_p as le_to_p,
   log1p_p as log1p_p,
+  log2_p as log2_p,
   log_p as log_p,
   logistic_p as logistic_p,
   lt_p as lt_p,